Add MySQL get connection utility. Rename petstore_util to test_util
