+++ test.config.ini
@@ -1,5 +1,8 @@
 [petstore]
 base_url = https://petstore.swagger.io/v2/
-pet_endpoint = pet/
-store_endpoint = store/
-user_endpoint = user/
+[SQL]
+host = localhost
+database = PythonAutomation
+user = cng
+password = cng1
+
